+++ docs/changelog.md
@@ -3,6 +3,5 @@
 Libraries:
 
 - [Android Allure Gradle Plugin](changelog_android_allure.md)
-- [Android Reports](changelog_android_reports.md)
+- [Test reports](changelog_reports.md)
 - [JUnit4 Rules](changelog_junit4_rules.md)
-- [Report Runtime](changelog_runtime.md)